Moves Reporter#calculate to own module
        Other than #start, all methods inside Reporter are private.
Extracting them out is going to make for easier testing and easier
reasoning.

        Also, added a callback to Reporter#start for use with future
testing.

        Right now the way that the methods other than storeCalculation
are attached to the module is kind of jank.
