+++ lib/reporter.js
@@ -5,10 +5,12 @@ var parse = require('url-parse');
 var request = require('request');
 var models = require('../models');
 var ReportSum = require('./report_sum');
+var store_calculations = require('./store_calculations');
 
 var Reporter = {
 
-  start: function (githubData, travisData) {
+  start: function (githubData, travisData, done) {
+    done = done || function() {};
     // temporary fix for numbered pull requests
     // TODO: handle this in continua11y script instead
     var pullRequest;
@@ -40,137 +42,8 @@ var Reporter = {
             branch: travisData.branch,
           }
         }).then(function () {
-          Reporter.calculate(githubData, travisData);
+          store_calculations(githubData, travisData);
         });
-      });
-    });
-  },
-
-  calculate: function (githubData, travisData, done) {
-    done = done || function () {};
-
-    let sums = new ReportSum(travisData.data).calculate();
-
-    let databaseSaves = Object.keys(travisData.data).map((url) => {
-      return function saveUrl(next) {
-        let parsedUrl = parse(url, true);
-        let data = travisData.data[url];
-        Reporter.saveURL(githubData, travisData, data, parsedUrl, next);
-      }
-    });
-
-    databaseSaves.push((next) => {
-      Reporter.saveCommit(githubData, travisData, sums, next);
-    });
-
-    databaseSaves.push((next) => {
-      if (travisData.pull_request === 'false') {
-        Reporter.saveRepo(githubData, travisData, sums, next);
-      }
-    });
-
-    databaseSaves.push((next) => {
-      Reporter.tellGitHub(githubData, travisData, sums, next);
-    });
-
-    async.parallel(databaseSaves, done);
-  },
-
-  saveURL: function (githubData, travisData, data, url, next) {
-    // record stats for each URL in the commit
-    models.Url
-      .create({
-        path: url.pathname,
-        commit: travisData.commit,
-        repo: githubData.id,
-        total: data.count.total,
-        error: data.count.error,
-        warning: data.count.warning,
-        notice: data.count.notice
-      })
-      .then(function() { next(); });
-  },
-
-  saveCommit: function (githubData, travisData, overall, done) {
-    done = done || function () {};
-    // records overall stats for this commit
-    models.Commit
-      .update({
-          total: overall.total,
-          error: overall.error,
-          warning: overall.warning,
-          notice: overall.notice,
-          latest: true
-        },{
-          where: {
-            commit: travisData.commit
-          }
-        })
-      .then(() => { done() });
-  },
-
-  saveRepo: function (githubData, travisData, overall, done) {
-    done = done || function () {};
-    // check if repo is tracked
-    models.Repo.upsert({
-      repo: githubData.id,
-      repoName: travisData.repository,
-      defaultBranch: githubData.default_branch,
-      total: overall.total,
-      error: overall.error,
-      warning: overall.warning,
-      notice: overall.notice
-    })
-    .then(() => { done(); });
-  },
-
-  tellGitHub: function (githubData, travisData, overall, done) {
-    done = done || function() {};
-    var context = '';
-    var message = '';
-    var state = '';
-    if (travisData.pull_request === 'false'){
-      context = "continuous-integration/continua11y/push";
-    } else {
-      context = "continuous-integration/continua11y/pull";
-    }
-    var firstCommit = travisData.commit.slice(0,6);
-    // var lastCommit = travisData.commit.slice(15,21);
-    models.Commit.findOne({
-      where: {
-        repo: githubData.id,
-        shortCommit: firstCommit
-      }
-    }).then(function (commit) {
-      var change = commit.error - overall.error;
-      if (change >= 0){
-        message = change+' fewer accessibility errors';
-        state = 'success';
-      } else {
-        change = Math.abs(change);
-        message = change+' more accessibility errors';
-        state = 'failure';
-      }
-      request.post({
-        uri: 'https://api.github.com/repos/'+travisData.repository+'/statuses/'+travisData.commit,
-        headers: {
-          'User-Agent': 'continua11y',
-          'Authorization': 'token '+process.env.GITHUB_TOKEN
-        },
-        json: true,
-        body: {
-          'state': state,
-          'target_url': 'https://continua11y.18f.gov/'+travisData.repository+'/'+travisData.commit,
-          'description': message,
-          'context': context
-        }
-      }, function (err, res, body){
-        if (err){
-          console.log(err);
-          done(err);
-        } else {
-          done();
-        }
       });
     });
   }
